+++ src/LibNoise.NET35/Renderer/Heightmap16Renderer.cs
@@ -77,7 +77,7 @@ namespace LibNoise.Renderer
         /// <param name="y"></param>
         /// <param name="source"></param>
         /// <param name="boundDiff"></param>
-        protected override void RenderHeight(int x, int y, float source, float boundDiff)
+        protected override void RenderHeight(int x, int y, double source, double boundDiff)
         {
             ushort elevation;
 
@@ -86,7 +86,7 @@ namespace LibNoise.Renderer
             else if (source >= _upperHeightBound)
                 elevation = ushort.MaxValue;
             else
-                elevation = (ushort) (((source - _lowerHeightBound)/boundDiff)*65535.0f);
+                elevation = (ushort) (((source - _lowerHeightBound)/boundDiff)*65535.0);
 
             _heightmap.SetValue(x, y, elevation);
         }
